Split sidecar tensor support into module
